Make sparse arrays not a extension
